uhuru/requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.6
django-audiofield==0.10.0
django-celery==3.3.1
django-cors-headers==3.4.0
djangorestframework==3.11.0
importlib-metadata==1.7.0
kombu==4.6.8
Pillow==7.2.0
pkg-resources==0.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
